complete first past using flask for intent parser
